requirements.txt
fastapi
uvicorn
pandas
scikit-learn
numpy
python-multipart
PyJWT
bcrypt